+++ shared/src/commonMain/kotlin/org/example/quiversync/domain/model/User.kt
@@ -8,8 +8,8 @@ data class User(
     val latitude: Double? = null,
     val longitude: Double? = null,
     val dateOfBirth: String? = null,
-    val heightCm: Int? = null,
-    val weightKg: Int? = null,
+    val heightCm: Double? = null,
+    val weightKg: Double? = null,
     val surfLevel: String? = null,
     val profilePicture: String? = null,
 )
